+++ src/components/AccountBalance/AccountBalance.tsx
@@ -1,16 +1,16 @@
 import { observer } from 'mobx-react-lite';
 import { Typography, TypographyProps } from '@cere-wallet/ui';
 
-import { useAccountStore } from '~/hooks';
+import { useBalanceStore } from '~/hooks';
 
 export type AccountBalanceProps = TypographyProps;
 
 export const AccountBalance = (props: AccountBalanceProps) => {
-  const { balanceStore } = useAccountStore();
+  const { balance, selectedToken } = useBalanceStore();
 
   return (
     <Typography {...props}>
-      {balanceStore.balance || 0} {balanceStore.selectedToken.displayName}
+      {balance || 0} {selectedToken.displayName}
     </Typography>
   );
 };