+++ src/stores/BalanceStore/BalanceStore.ts
@@ -1,27 +1,29 @@
 import { makeAutoObservable } from 'mobx';
 import { getTokenConfig } from '@cere-wallet/wallet-engine';
 
-import { AccountAssetStore, Asset } from './AccountAssetStore';
+import { Asset } from '../types';
+import { AssetStore } from '../AssetStore';
 
-export class AccountBalanceStore {
+export class BalanceStore {
   selectedToken: Omit<Asset, 'balance'>;
 
-  constructor(private assets: AccountAssetStore) {
+  constructor(private assetStore: AssetStore) {
     makeAutoObservable(this);
 
     const token = getTokenConfig();
     this.selectedToken = {
       displayName: token.symbol,
-      ticker: token.symbol.toLocaleLowerCase(),
+      ticker: token.symbol,
+      network: 'Polygon',
     };
   }
 
   get nativeBalance() {
-    return this.assets.nativeToken?.balance;
+    return this.assetStore.nativeToken?.balance;
   }
 
   get balance() {
-    const selectedToken = this.assets.list.find(({ ticker }) => this.selectedToken.ticker === ticker);
+    const selectedToken = this.assetStore.assets.find(({ ticker }) => this.selectedToken.ticker === ticker);
 
     return selectedToken?.balance;
   }