+++ src/stores/types.ts
@@ -1,3 +1,4 @@
+import { ChainConfig } from '@cere-wallet/wallet-engine';
 import { ethers } from 'ethers';
 
 export type PriceData = {
@@ -6,8 +7,24 @@ export type PriceData = {
   equalsTo?: Omit<PriceData, 'equalsTo'>;
 };
 
+export type Account = {
+  address: string;
+  privateKey: string;
+  email: string;
+  avatar?: string;
+};
+
+export type Asset = {
+  ticker: string;
+  displayName: string;
+  network: string;
+  balance?: number;
+};
+
 export type Provider = ethers.providers.JsonRpcProvider;
 export type Wallet = {
   readonly instanceId: string;
-  readonly provider: Provider | null;
+  readonly network?: ChainConfig;
+  readonly provider?: Provider;
+  readonly account?: Account;
 };