+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### vNext
 
--
+- [DAV-3036] Wallet asset overview widget
+- [DAV-3119] Assets balance real-time update
 
 ### v1.2.0